core/src/main/java/alexey/grizly/com/users/services/impl/AuthServiceImpl.java
package alexey.grizly.com.users.services.impl;

import alexey.grizly.com.users.dtos.request.AuthRequestDto;
import alexey.grizly.com.users.services.AuthService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.authentication.AuthenticationManager;
import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
import org.springframework.security.core.Authentication;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.stereotype.Service;

@Service
public class AuthServiceImpl implements AuthService {
    private final AuthenticationManager authenticationManager;

    @Autowired
    public AuthServiceImpl(AuthenticationManager authenticationManager) {
        this.authenticationManager = authenticationManager;
    }
    @Override
    public UserDetails authentication(AuthRequestDto dto){
        Authentication authentication = authenticationManager
                .authenticate(new UsernamePasswordAuthenticationToken(dto.getEmailOrPhone(), dto.getPassword()));
        return (UserDetails) authentication.getPrincipal();
    }
}
